Add script for populating public cache (#21)
Added a script to make it easier to publish to our public cache. Needed this for our staging environment.

The deno images now have official mirrors, so I moved to that. The old ones no longer exist.

I didn't set up gcloud access on the Airplane task, so for now this task is only run locally.