Handle init for existing tasks (#362)
Properly handles `init` for existing tasks. Also does not generate a generic entrypoint for existing tasks, but informs the user where the entrypoint is set. This is probably not a _great_ solution--if you run `airplane init --dev --from hello_world hello_world.py`, the file argument is completely ignored. Maybe the `entrypoint` should be overwritten in that case?

Also fixes a bug where if a definition file was supplied to `init` it would not get used & instead default to `<slug>.task.<format>`.

Depends on https://github.com/airplanedev/lib/pull/24. Resolves AIR-2460.