Break out build config as a separate concept (#51)
Differentiate somewhat between "options that are part of the task definition" & "options that are part of the build config". This uses task KindOptions as a base, & merges it with BuildConfig that is tracked separately. It enables us to tweak build-related options without actually changing the task definition (e.g., workdir for Node, or rewriting the entrypoint to be relative to the task root).

Downstream services should call GetBuildConfig instead of GetKindAndOptions for build options.