Always ask for an entrypoint (#416)
Previously we were only prompting the user for an entrypoint if we didn't have one to use. Now we always prompt the user for an entrypoint. In order of decreasing priority, the default entrypoint is (a) the file provided to `airplane init`, if the file is not a task definition file, (b) the entrypoint specified on the existing task, or (c) `<slug>.<ext>` based on the task kind. For usages of `airplane init --from <slug>`, this gives users the ability to double-check the entrypoint; it also means that the given file is not ignored for usages of `airplane init --from <slug> <entrypoint>`.

The entrypoint prompt was tweaked so as to not imply that we're always going to create the entrypoint.

Also resolves AIR-3190.