Support deploying tasks to new environments (#412)
To support deploying tasks into environments where they are not already deployed, we need to update our deployment logic to not depend on a full `Task` struct. This PR updates `airplanedev/lib` to conform to the new APIs from: https://github.com/airplanedev/lib/pull/75

The functional change is that when deploying a task in an environment that it hasn't been deployed to, you no longer are prompted to create the task:

```sh
# Before (prompts to create task)
❯ lairlocal deploy --dev --debug --env=stage ./env_deployments/env_deployments.task.yaml
[...]
[debug] performing request GET https://api.airplane.so:5000/v0/tasks/get?envSlug=stage&slug=env_deployments
? Task with slug env_deployments does not exist. Would you like to create a new task? (Y/n)

# After (task already exists, just not in this env, continues to deployment)
❯ lairlocal deploy --dev --debug --env=stage ./env_deployments/env_deployments.task.yaml
[...]
[debug] performing request GET https://api.airplane.so:5000/v0/tasks/getMetadata?slug=env_deployments
Deploying 1 task:
[...]
```

In a follow-up PR, I'll adjust our deployments logic to actually deploy into the new environment (currently it just deploys to the default environment).