Respect working directory of custom Dockerfile (#81)
If the user specifies a custom Dockerfile when running a shell task, run the shell script in the same directory as the custom Dockerfile. Currently, we copy the shell script into /airplane - if a user wants to access files created by their custom Dockerfile, it isn't obvious that they are now inside the /airplane directory. This is more intuitive - our shim no longer changes the working directory and it's as if the shell script is run at the end of the custom Dockerfile.

https://linear.app/airplane/issue/AIR-2842/shell-builder-should-work-with-different-workdir