Add TransactionMode to SQL tasks (#98)
## Description

New SQL tasks get created with a default TransactionMode of auto. Initializing SQL tasks without a TransactionMode results in no transactionMode entry in the definition file. Resolves AIR-2896.

## Test plan

Deploy a SQL task with a transaction mode. Initialize a SQL task from an existing task with a transaction mode, and without a transaction mode.