Add configs to task definition yaml (#104)
## Description
This PR adds a `Configs` field to the task definition YAML for REST and SQL tasks, similar to how there is an `EnvVars` field for code tasks. It also adds a `Configs` field to a couple structs so that the CLI can deploy a task with config attachments. 

## Test plan
Unit tests