Add build args to Dockerfile for node and python (#111)
## Description
In order to use build args in a Dockerfile, the arg must be declared in the Dockerfile. We are currently passing all user env vars as build args, but not declaring these args in the Dockerfile, so they cannot be used.

This will require one more Airport PR where we pass build arg keys when generating the Dockerfiles.

## Test plan
Wrote tests for python and node. Will test manually once the Airport PR is out

AIR-3569