Discover tasks by script if exists a defn with the same name (#109)
## Description
If a user runs `airplane deploy my_task.ts` and there is a file `my_task.task.yaml` in the same directory, we will deploy `my_task.task.yaml`. This only works if the task script is in the same directory as the task yaml.

## Test plan
Unit test. Plus I will test using the CLI when this is merged

FIXES AIR-3325