Send request envSlug as query param (#462)
## Description
We want to start sending envslug information as a query param in the external api. This change makes the cli start sending the env info in as a query param in addition to what it does now. Next step will be making the "breaking" change in api-server (for those that want to use environments in their cli), and then we can cleanup this cli.

## Test plan
...