Resolve issues with importing latest Node SDK beta (#114)
## Description

1. Resolves an issue where users could not install their own version of the `airplane` SDK (e.g. the new `0.2.0` beta). Resolves: https://linear.app/airplane/issue/AIR-3585/users-cannot-override-node-shim-dependencies
2. Resolves an issue where the `airplane` SDK, which depends on `node-fetch` starting in `0.2.0`, was getting marked as `external` and therefore not converted to CommonJS.

Due to (1), this is a breaking change for any Node tasks that (1) depend on a `0.0.x` version of the Node SDK and (2) use the output methods (`setOutput`, `appendOutput`) introduced in `0.1.x`. IOW, these tasks are implicitly relying on this bug to give them the newer version of the SDK. The fix is to update your `package.json` to reference the version of the SDK you want, e.g. `~0.1.2`. Many of our examples had this issue and I went through and made that change. I'm considering adding a safeguard, where we check if the user is requesting a `0.0.x` version of `airplane` and if so, continue to install the newer shim version of airplane. However, (1) we would need to parse node semver syntax (probably easier to just regex it?) and (2) as we bump the version of the SDK that the shim uses, this could cause issues for these customers. Because of (2) and because this only impacts rebuilds, I lean towards one-off helping customers who hit this issue.

Resolves AIR-3585

## Test plan

Ran locally and confirmed the built image runs successfully. Tested with a few variants of JS tasks (including w/ and w/out a dep on `airplane`).