Upgrade lib and clean up go.mod file (#473)
## Description
This change updates the lib version used by our cli to pick up the changes made in https://github.com/airplanedev/lib/pull/120. It also cleans up the `go.mod` file so that direct dependencies are in the first `require (...)` block and indirect dependencies are in the second. This is how go does it for new projects, but if the go.mod was created with an earlier version of go, this file can have multiple require blocks over time.

cc: @calufornia 

## Test plan
Testing completed successfully via running locally, verifying that I could deploy tasks with both `runtime: durable` and `runtime` unset.