Support building an app (#118)
## Description
Support building a vite app. A few things here are very POC and will have to be revisited.

For now, we only support vite and assume that the app has been bootstrapped with vite and has a `vite.config.ts`.

We build the app and store the build in the `/dist` folder. After the image is built, we will upload the contents of the `dist` folder to gcs to be served.

## Test plan
Wrote a unit test to ensure build succeeds.

FIXES AIR-3623