Prompt user to confirm deploy if there's a diff (#490)
Now that we are writing a task defn diff, I think it makes sense to prompt the user for confirmation before deploying. This is what I would expect with a system like this. This basically creates a "preview" mode where you can preview changes and then apply them.

If there is no task defn diff or the task is new, we don't prompt for confirmation.