Remove unnecessary call to list resources (#136)
## Description

I noticed this when making some changes for resource encryption - we define a mapping from name -> resource, populate it with the result of `/resources/list`, but then don't do anything with this map. This removes the unnecessary API call to list resources.

## Test plan

Tested a CLI deploy of a SQL task just to be sure, but this change should be safe since we're just removing unused code.