Fix npm/yarn workspaces deploy & dev (#132)
We currently have a few bugs in our implementation of npm/yarn workspaces

When deploying, we only gather dependencies to mark as external from the root package.json. This means that any packages declared in a workspaces's package.json (generally where you declare all of your dependencies) are ignored and not marked external. This causes any dependencies that cannot be bundled to break the deployment.

When devving, we gather dependencies from the workspace only and not the root package.json (opposite as deploying :confounded: ). Moreover, we consider any sibling workspaces as dependencies and mark them external which causes us not to build them - this causes dev to not work if a workspace depends on another workspace (it often does).


Instead, what we have to do is:
1. Gather deps from all package.jsons of all of the workspaces (including the root) to mark as external... but
2. Do NOT mark any imported, sibling workspaces as external so that they are bundled.

To accomplish 2, we use the command `yarn workspaces info` which prints out the individual workspaces in a project and ensure these are not marked external so they can be built.

A few comments:
1. We require that users have `yarn` so we can run `yarn workspaces info`. We will install this in our own bundler so that we have it for deployments. We could remove this requirement, but it would mean we have to do the work of `yarn workspaces info` which seems pretty annoying
2. I still want to discuss where we want to move in terms of bundling in general.

Please let me know if we want to discuss any of this!