Make custom Temporal activities optional (#155)
## Description

Right now, we assume that a task's custom activities exist in a file called `activities.ts` or `activities.js` at the root of their project. However, a task may not have custom activities (if it is purely using the Node SDK) - this PR makes it so that we don't encounter an import error if neither of these files exist by creating an empty `activities.js` file, similar to how we create an empty `package.json` if the task does not contain one. The activities that are registered with the worker is the union of the activities exported by the Node SDK and the task's custom activities.

cc @colinking @yolken-airplane 

Resolves AIR-3973

## Test plan

Tested locally by deploying and running a task with & without `activities.ts`.