Monkey patch process.env in temporal workflow runtime (#159)
## Description

Adds an activity that loads in environment variables from the main node process. We then monkey patch `process.env` so that the user can continue to reference environment variables in their code in the same way as standard tasks. This `process.env` object will also contain some airplane-specific values that are used to determine which execute method in the SDK to call, env, etc.

## Test plan

Tested by deploying/running a workflow task locally without specifying `runtime: "workflow"` in the `ExecuteOptions`.