Don't log during workflow replay (#176)
## Description

Sets `callDuringReplay` to `false` for the logger sink. This was previously required because [there was a bug](https://github.com/temporalio/sdk-typescript/issues/667) that caused the entire workflow (even steps that had not executed yet) to be flagged as replaying, which meant that any steps after a workflow interruption would not be logged if the field was not set to true. However, this introduced another bug - by enabling the flag, duplicate logs would occur for all workflow steps that occurred before the interrupt. Upgrading the airplane SDK version to one that uses the latest temporal typescript SDK lets us ensure that each log occurs exactly once.

Also contains various changes that are required as a result of breaking changes in the latest typescript SDK: https://github.com/temporalio/sdk-typescript/blob/main/CHANGELOG.md

cc @yolken-airplane 

## Test plan
Interrupted a workflow by delete the runner pods during a workflow execution, and verified that the task was still able to complete successfully.