Patch extra files needed for views deployment (#179)
## Description
Still have a bit of cleanup to do, but going to put up this PR first - basically the equivalent of the CLI dev change but for deploy. Needs updated PRs in both `airport` and `cli` that I'm going to put up soon

## Test plan
Ran views init and views deploy using CLI (PR incoming) and verified that it worked.