Add ability to reference local resources for airplane dev (#530)
## Description

This PR adds the notion of an airplane dev config file, which will be used to store local resource configuration. The dev config will store a mapping from slug to resource configuration - this is because the task definition file stores resource attachments in the form of a map between alias to slug, and so combining the two we have a mapping from alias to resource configuration, which is what is required by our SDK. We add logic to encode resource configs into an `AIRPLANE_RESOURCES` env var, which will then be used by the Node SDK to get a task's resource attachments and builtins.

This PR also copies over many of the builtins structs - these should eventually go into their own repo but are included here for now. 

## Test plan
Tested with the below dev config and verified that resources were serialized into the `AIRPLANE_RESOURCES` env var properly:
```
env:
  API_KEY: asdf
  IMAGE: 1234
resources:
  db:
    kind: postgres
    slug: db
    username: admin
    password: airplane
  rest:
    kind: rest
    slug: rest
    baseURL: "localhost:1234"
```