Add Inter font link (#531)
## Description

The Inter font package is too large. The inter package dependency is getting removed in https://github.com/airplanedev/views/pull/86

Rather than downloading it, add a link to the hosted font

Resolves AIR-4255

## Test plan

Built CLI and do airplane dev on a view. Verify that the `<head>` component has the inter link