Check if we should use yarn for views init (#532)
## Description
We try to use yarn if user has a `yarn.lock` or if they have yarn installed unless the user is creating their view in a directory with a `package-lock.json`

RFC because I'm not sure if we want to do this or just default to using npm all the time if neither lockfile exists even if they have yarn installed. Currently tasks will default to npm if no lockfile is found.

## Test plan
Uninstalled yarn, init in a fresh directory and it uses npm
Re-installed yarn, init in a fresh directory and it uses yarn

With yarn installed, init in a nested directory with a package-lock.json and it uses npm

Is there an easy way to write a test for this?