Add airplane server command (#538)
## Description

Adds the `airplane server` command, which is used to start the local dev server in the foreground. The command takes in an optional argument that configures the folder where tasks and views should be discovered in. Also adds a new `/dev/` subrouter, which currently just contains a ping handler - these endpoints will eventually called by the airplane previewer. We leave the "short-running" airplane server in `airplane dev` for now, but eventually we'll want to replace `airplane dev` with a "long-running" server.

## Test plan
```
aplocal server ~/airplane/node-sdk
Discovering tasks and views...
Found 2 tasks and 0 views:
- Node task
- Workflow task
```

```
curl "localhost:7190/dev/ping"
ok%
```