Add custom json marshaller for resource attachments (#199)
## Description

Something I noticed when testing https://github.com/airplanedev/cli/pull/540 - we currently have a custom JSON unmarshaler for resource attachments, but not a marshaler. This causes resources attachments to by default get (JSON) marshaled as
```
...
"resources": {
  "attachments": {
    "db": "demo_db"
  }
}
...
```
This subsequently causes unmarshaling to fail since a map of maps is unexpected. In actuality, resource attachments should be marshaled as 
```
...
"resources": {
  "db": "demo_db"
}
...
```
This isn't a problem today because there _is_ a `MarshalYAML` method that properly marshals the definition into yaml for deploys (we use it for json too) - this is mainly a problem in unit tests that don't use a yaml marshaler.

## Test plan
Tested by deploying a task with resource attachments locally - also unit tests in above PR.