Require task slug in createRuns endpoint (#575)
## Description
Fixes bug where when a run is created before task execution is finished (to stream logs), that run is not associated with the task. We now require createRun to have a task slug.

Also only adding a run to a task's runHistory if it's not already there. Since that can happen at create run or when executing a task (if the run hasn't already been created)... - open to suggestions here but it feels like we'll want to move to ExecuteTask returning a runID immediately and executing asynchronously 

Other fixes to local runs store:
- adds a lock to the local runs store
- create/update populate the RunID and task name inside the run object, for convenience so the client doesn't need to add this to the LocalRun object every time

## Test plan
Tested that getting a task, executing a task, and getting logs for a run still works
verified the run is associated with the task