Fix yarn workspace dependency installation (#221)
## Description

This PR contains a couple changes related to yarn workspaces that are required for https://github.com/airplanedev/node-sdk/pull/96 to work. In particular, when generating the shim `package.json`, look to see if any `package.json` between the root `package.json` and the script's `package.json` satisfies the `airplane` packages - if so, no need to install. Previously, we would only look at the root `package.json`; when using workspaces, dependencies are often declared in the workspace `package.json` file instead of the root, and so we should check there too.

## Test plan
Tested locally by deploying and running the tasks included in https://github.com/airplanedev/node-sdk/pull/96