Add subcommands to modify dev config file (#578)
## Description
Adds `airplane dev config set-resource`, `delete-resource`, `set-env`, and `delete-env` commands that modify the dev config file. These will eventually be modified through the IDE, but for power users or those wanting to modify the file through the CLI, they can use these commands. If a dev config file does not exist, the user is prompted to create one with the name `airplane.dev.yaml`.

## Test plan
Tested operation of the above subcommands.