Add kind configs to lib (#230)
## Description

This PR adds all the kind configs to the lib repo, which is necessary because we're changing the format of resources in the dev config file to use the api version of the resource, as opposed to the builtins versions of resources (i.e. `/pkg/resources/kinds`). We also include most of the interface methods of the kind configs since they give us the ability to convert the kind configs into a builtin resources (which we need for local dev), validation, etc.

## Test plan

n/a