Add ID method to external resource interface (#231)
## Description

The local dev server needs to use the resource id in order to figure out which resource to encode in the `AIRPLANE_RESOURCES` env var, which is used by the builtins binary. This is because the SDK sends the resource ID ([example](https://github.com/airplanedev/node-sdk/blob/4a617af844238aecec947617f8a7e7c8117e34e2/packages/airplane/internal/builtins/rest.ts#L44)) in the request to `/v0/tasks/execute` - in the local dev server, we'll compare the resource id in the request with the ids of the resources in the dev config file. This PR exposes an `ID()` method that gets the ID field from the base resource.

Note: This change affects the external resource structs - we'll be using the api resource structs temporarily (which have an ID field already exported), but we'll still need this change once we go back to using external resources.

## Test plan

Tested locally with corresponding CLI changes.