Add support for displays during local dev (#579)
## Description

Adds support for seeing displays when locally developing on a task. The displays are stored in the run store where they can be accessed by the UI. They're also logged out for the existing `airplane dev` usage, but we can remove that once we fully migrate all local dev to the server.

Resolves AIR-4485

## Test plan

Run a task locally that produced displays. The displays should render in the UI when running https://github.com/airplanedev/web/pull/1639