Fix log store synchronization (#593)
## Description

This PR fixes some issues with thread-safety when streaming logs to the previewer. Most notably, logs were dropped before due to a race condition where the executor was (attempting to) send logs through the log channel while the SSE handler was already processing a log - the log channel is an unbuffered channel (since we don't know how many logs the user's code will emit), and so during this processing, the executor couldn't send a log through the channel and would fallback to the default behavior of just appending it to the log store. 

We've fixed the above issue and refactored some of the logging logic out of the dev server and executor into a `LogBroker`. The log broker handles writing/sending logs and registering a set of `LogWatcher`s that receive from the broker. The broker handles the logic of fanning out log messages to watchers and also handles synchronization, which makes the implementation of the dev server and executor simpler.

cc @kellymtrinh 

## Test plan
Tested the following scenarios with the previewer with the following code, both with and without the commented-out line:
```
export default async function(params) {
    for (let i = 0; i < 100; i++) {
        console.log(i);
        // await new Promise(resolve => setTimeout(resolve, 100));
    }
}
```

1. Executed a task and viewed until completion.
2. Executed a task, navigated away, navigated back, viewed until completion.
3. Executed a task, navigated away, waited for completion, navigated back.