Deprecate node v12 and 15, and set v18 as default (#124)
## Description
We've deprecated support for Node v12 and v15. This PR removes references to these versions from task definitions, and also sets "18" as the default version, which is required for workflow tasks and matches the default version when creating a node task through the UI.


## Test plan
Initialized a new node task locally and verified the version was "18", attempted to change the node version to "12" and deploy, and verified that schema validation failed.