Add local prompts (#609)
## Description
Adds support for local prompts in the previewer. This implements the same internal/external prompt endpoints used in the SDK and web. (It's split over ext and int api to match what we do in airport so we can reuse the hooks in web.)

I decided to store the prompts on the run itself instead of keeping track of them separately (since updating/submitting prompts seems like the most complicated op here), but open to feedback.

## Test plan
-  <img width="674" alt="Screen Shot 2022-08-25 at 2 28 39 PM" src="https://user-images.githubusercontent.com/8205327/186772272-3d1bdd5d-b80c-4f8c-94f2-aa6896b2668f.png">

- <img width="670" alt="Screen Shot 2022-08-25 at 2 25 04 PM" src="https://user-images.githubusercontent.com/8205327/186771776-ae7ca8b2-f8b2-4a3f-9b8d-b851a048e2ae.png">