Create dev config file on-demand and generate resource slug (#616)
## Description

If the user doesn't pass in a dev config file, creates one in the current directory on the first "write" operation (e.g. creating a new resource) - previously, the dev server would panic if attempting to create a resource when the dev config file didn't exist. Also generates a resource slug if one isn't provided to the resource creation endpoint - users should ideally be setting this from the UI to match what they have in their resource attachments, but we'll fall back to auto-generation if one isn't provided.

## Test plan
Tested locally by creating a new resource with no `airplane.dev.yaml` present, and verified that the dev config file was created with a slug generated from the name.