Fix loading task-based views in editor (#620)
## Description
Previously, we would override the root api client's host so that any API requests made by the discoverers would get routed to the local dev server. However, this also meant that any task execution requests from a view (which uses the root api client) would also get routed to the local dev server, and would fail since we don't support local view -> local task execution yet. We fix this so that the view will hit whatever host is specified in the dev command (usually just api.airplane.dev), and not the local dev server. This makes loading views from the editor equivalent to `airplane views dev`.

## Test plan
Tested by loading a view backed by a task, and verified that the task executed successfully and the table was populated with the task's output.