Return task along with run in local dev server (#624)
## Description

Returns the task along with the run in the call to `/runs/get`, similar to what we do in our production API, which allows to remove the `/tasks/get` call in the frontend when displaying the run details. Also adds `taskID`, which is equivalent to the task slug for local tasks.

## Test plan
Tested executing a task locally and verified run details were as expected.