Don't automatically set @airplane/workflow-runtime version if set in task dependencies (#259)
## Description

We previously added logic that would align the versions of `airplane` and `@airplane/workflow-runtime` if the former was included in the task's dependencies. This is usually the behavior we want, but we don't want to do this in the case that the task's dependencies already include `@airplane/workflow-runtime`, which causes the workflow runtime package to get installed twice (once from the shim dependencies, and another time from the task's dependencies). For example, this is the case in our [example workflow task](https://github.com/airplanedev/js-sdk/blob/01bf802ad329ccc13519a4898257aa7a30035259/examples/workflow/package.json#L5), which utilizes yarn workspaces to use the local version of the workflow-runtime package.

## Test plan

Tested by deploying https://github.com/airplanedev/js-sdk/blob/main/examples/workflow/main.airplane.ts