Cache editor version response  (#642)
## Description
Github heavily rate limits the endpoint used to check if the CLI is running the latest version. We can cache this in mem for the duration of the server right now. When we add hot reloading we should periodically refresh this information. (And use a better lib for storing runs too that imposes limits on max size, etc.)

The version endpoint was periodically pinged from the UI to ensure the server is still running, and checked that the version is latest. 

Deleting the ping endpoint since that's now unused.

## Test plan
- Curl endpoint still works
``` 
❯ ~ curl http://localhost:4000/dev/version

{"status":"ok","version":"\u003cunknown\u003e","isLatest":true}
```

- Can still execute a task and navigate around in the editor.