Handle ExportResource in Resource objects (#269)
## Description

Maddeningly, we have both `api.Resource` as well as `kind_configs.InternalResource`. I'm leaving these alone for now, since eventually we'll get rid of the whole `kind_configs` package.

This adds `ExportResource` (JSON-serialized as `resource`) to both `api.Resource` & `kind_configs.InternalResource`. It also adds `ExportResource` to conversion logic. Note that we now need a custom `UnmarshalJSON` for anything that includes `stdresource.Resource`, because it's an interface & not an actual type. This mirrors the changes made in https://github.com/airplanedev/airport/pull/3173.

## Test plan

Tested with corresponding changes in `cli`.