Bump JS SDK used by JS shims (#272)
## Description

Updates `lib` to use the latest version of the JS SDK (`v0.2.0`). Most of the changes are from the test fixtures, but more importantly this changes the `defaultSDKVersion` to `0.2.0` so that shims will install the newer SDK automatically.

Since we removed `airplane.output` in the most recent version of the SDK, customers who were relying on `airplane` getting auto-installed for their tasks and that still use the legacy output SDK methods will find that their tasks do not build anymore. This does not affect existing tasks.

## Test plan

Deploy a task with the latest lib.