Remove dummy (#279)
## Description

Remove need to shim airplane view exports. The reason why the views library can't be run in a node environment is because it has some code that assumes it's in a browser environment. We're going to find the same thing true with other packages that our users use. For a more generic fix, use the `jsdom` package in parser.js to apply a fake dom to the user's view - this should allow code that expects to be run in a browser to run successfully.

## Test plan

I wrote some unit tests for this. We're now actually testing real views code. This required me to push up the node_modules of the test views (otherwise we'd have to run yarn install on every fixture which is annoying. we could do it but it'd add a bunch of extra time and it works just pushing up node_modules).