Allow local views to execute local tasks (#667)
## Description

Adds the ability to execute a local task when developing a view in the editor. We introduce a `--local` flag to the `airplane dev` command, which defaults to true and indicates that any task referenced by a view should be executed locally instead of in one of the user's team's environments. If `--local=false`, the task will be executed in the user's default environment (usually production), or if an `--env` is specified, then the designated environment.

## Test plan
Tested by creating a view with a task-backed component and verifying that the view called the local task and the component's data was populated correctly.