Support slack resource in editor (#684)
## Description
Adds support for Slack builtins from the editor - Slack is a special builtin with a set id and slug, and so must be handled separately. We also "include" the Slack resource by default so that the user does not need to explicitly set the `--env` flag - Slack resources cannot be created in local dev, and so the `--env` flag would be redundant.

## Test plan
Tested by executing a task that called a Slack builtin, and verifying that it worked even without explicitly setting the `--env` flag.