Allow remote execution of tasks (#685)
## Description

Similar to resources, if the user supplies a fallback environment using `--env` and the task is not registered locally, we will attempt to execute the task in the user's given environment. We do this by essentially making the local dev server a "proxy" to our production API if a task has not been found locally, i.e. forwarding the `/v0/tasks/execute` in the task does not exist in the task config map.

Also removes the `/dev/tasks` endpoint, which is unused, and updates some of the language around dev server startup.

## Test plan

Tested locally by running a task locally that called another task that was not registered locally, and verified that the remote task ran properly and we were able to get the outputs from it.