Check if slack resource exists before merging with local resources (#695)
## Description
Currently, we attempt to automatically add a team's Slack resource for convenience, so they don't have to specify a `--env` fallback. If a team hasn't configured Slack, however, this leads to a confusing error message to the user about the request to `/v0/resources/get` failing. Similar to demo DB, we should only add slack to the editor's merged (local + remote) resources map if it exists remotely.

The user also receives a better error message when they call a slack builtin but haven't configured Slack: `Your team has not configured Slack. Please visit https://app.airplane.so:5000/settings/team to authorize Slack to perform actions in your workspace.`

## Test plan
Tested in conjunction with API changes to return a 404 if a team has not configured Slack.