Normalize definition in discoverer (#319)
## Description

I looked into adding this as a necessary step of `Unmarshal`, but `Unmarshal` is called (via `taskdir.ReadDefinition`) in so many places that don't actually need normalization (read the definition to get the slug, read the definition to get the entrypoint) that it would have been way more invasive to add it to `Unmarshal`. It would have also added a ton more API requests to a lot of commands, & (to my knowledge) we don't have any caching in the APIClient. (Maybe we should?) I added it as an optional step instead, which means that there is still handling for `Resource` being the name of a resource rather than a slug in other places, since there's no guarantee that the caller has called `Normalize` before calling (say) `GetUpdateTaskRequest`.

Resolves TSK-735.

## Test plan

Tested locally.